GenAIDemo/requirements.txt
SQLAlchemy
langchain
langchain-community
langchain-openai
python-dotenv
pyodbc
openai